fix(SyncWorker): changed fetching strategy for addr, tx and initial boot
