+++ src/EVENTS.js
@@ -11,7 +11,7 @@ module.exports = {
   BLOCKHEADER: 'blockheader',
   FETCHED_ADDRESS: 'FETCHED/ADDRESS',
   ERROR_UPDATE_ADDRESS: 'ERROR/UPDATE_ADDRESS',
-  FETCHED_TRANSACTIONS: 'FETCHED/TRANSACTIONS',
+  FETCHED_TRANSACTION: 'FETCHED/TRANSACTION',
   FETCHED_UNCONFIRMED_TRANSACTION: 'FETCHED/UNCONFIRMED_TRANSACTION',
   FETCHED_CONFIRMED_TRANSACTION: 'FETCHED/CONFIRMED_TRANSACTION',
   GENERATED_ADDRESS: 'GENERATED_ADDRESS',
